Don't use assembler for x86-64.
This reverts a change I thought I had to make in a recent upgrade, and
chooses the safer build fix instead. I note that the Linux kernel seems
to disable this too, but the proximate cause for me disabling this is
that zstd v1.5.5 causes CFI failures in logd with this enabled.

Test: treehugger
Change-Id: I8ae748ec36737685a1771f1f779b9896559dc258